edit and saving form info in local storage

+++ client/src/components/addGenerator/addImages/ImagesList.jsx
@@ -7,16 +7,17 @@ import {
 } from '@mui/material';
 import React from 'react';
 import { useValue } from '../../../context/ContextProvider';
-import deleteFile from '../../../firebase/DeleteFile';
+import deleteFile from '../../../firebase/deleteFile';
 
 const ImagesList = () => {
   const {
-    state: { images, currentUser },
+    state: { images, currentUser, updatedGenerator },
     dispatch,
   } = useValue();
 
   const handleDelete = async (image) => {
     dispatch({ type: 'DELETE_IMAGE', payload: image });
+    if (updatedGenerator) return dispatch({ type: 'UPDATE_DELETED_IMAGES', payload: [image] })
     const imageName = image?.split(`${currentUser?.id}%2F`)[1]?.split('?')[0];
     try {
       await deleteFile(`generators/${currentUser?.id}/${imageName}`);
